Move core proxy logic into lib.rs; accept ports as args
